modified transform.js to accept both palette and non-palette images and invert colors
